Added ability to check defending effectiveness
Type matchup logic in original version doesn't calculate correctly for two types. This update fixes the calculation. It lets the user see effectiveness of a specified type vs the selected Pokémon. However, it removes the "weak vs" and "strong vs" sections of the output until the calculations for these sections are fixed for dual types.